[BE] feat: Add session base auth in posts api

+++ src/controllers/postController.js
@@ -5,7 +5,8 @@ const { postDao } = require('../models');
 const { timeUtils, pagination } = require('../utils');
 
 // TODO: 게시글 작성자 임시 (JWT 구현중)
-const tempPostsOwnerId = 1;
+// const tempPostsOwnerId = 1;
+
 const updateAllowField = ['title', 'contents', 'thumbnail'];  // 수정 가능 부분
 
 // 5 게시글 씩 전송 (쿼리 요청 -> 5 +1 hasNext 필드)
@@ -47,6 +48,8 @@ const createPosts = (req, res, next) => {
     return res.status(400).json({ message: '필수 필드 누락' });
   }
 
+  const nowMember = req.member;  // session 기반
+
   const newPosts = {
     title,
     contents,
@@ -55,7 +58,8 @@ const createPosts = (req, res, next) => {
     commentsCount: 0,
     hitsCount: 0,
     likesCount: 0,
-    ownerId: tempPostsOwnerId,  // 임시 아이디
+    // ownerId: tempPostsOwnerId,  // 임시 아이디
+    ownerId: nowMember.memberId,
     isVisible: true,
   };
   postDao.save(newPosts);
@@ -70,6 +74,11 @@ const updatePostsDetails = (req, res, next) => {
   const findPosts = postDao.findById(postsId);
   if (!findPosts) {
     return res.status(404).json({ message: '일치하는 게시글 없음' });
+  }
+
+  const nowMember = req.member;  // session 기반
+  if (findPosts.ownerId !== nowMember.memberId) {
+    return res.status(403).json({ message: '권한이 없으' });
   }
 
   Object.keys(req.body).forEach((key) => {
@@ -89,6 +98,11 @@ const deletePosts = (req, res, next) => {
   const findPosts = postDao.findById(postsId);
   if (!findPosts) {
     return res.status(404).json({ message: '존재하지 않는 게시글' });
+  }
+
+  const nowMember = req.member;  // session 기반
+  if (findPosts.ownerId !== nowMember.memberId) {
+    return res.status(403).json({ message: '권한이 없으' });
   }
 
   postDao.deleteById(postsId);
